refactor: update import paths to use new utils structure for logger and fetch utilities
