+++ apps/electron/src/main/domain/agent/deployed/deployed-agent-service.ts
@@ -4,7 +4,7 @@ import {
 } from "@mcp_router/shared";
 import { getDeployedAgentRepository } from "../../../infrastructure/database";
 import { DeployedAgent } from "./deployed-agent";
-import { logError, logInfo } from "@/main/infrastructure/logger/logger";
+import { logError, logInfo } from "@/main/utils/logger";
 import { SingletonService } from "@/main/application/core/singleton-service";
 
 /**
