+++ apps/electron/src/main/domain/agent/development/development-agent-service.ts
@@ -8,7 +8,7 @@ import { getAgentRepository } from "../../../infrastructure/database";
 import { getServerService } from "../../mcp-core/server/server-service";
 import { DevelopmentAgent } from "./development-agent";
 import { v4 as uuidv4 } from "uuid";
-import { logError, logInfo } from "@/main/infrastructure/logger/logger";
+import { logError, logInfo } from "@/main/utils/logger";
 import { SingletonService } from "@/main/application/core/singleton-service";
 
 /**
