+++ apps/electron/src/main/domain/agent/shared/agent-base.ts
@@ -4,7 +4,7 @@ import {
   MCPAgentToolPermission,
   BaseAgentInfo,
 } from "@mcp_router/shared";
-import { logError, logInfo } from "@/main/infrastructure/logger/logger";
+import { logError, logInfo } from "@/main/utils/logger";
 
 /**
  * エージェントの共通ベースクラス
