+++ apps/electron/src/main/utils/logger.ts
